requirements.txt
accelerate>=0.33.0
bitsandbytes>=0.43.1
datasets>=2.20.0
evaluate>=0.4.2
huggingface-hub>=0.23.4
langdetect>=1.0.9
numpy
peft>=0.12.0
pillow>=10.3.0
pymupdf>=1.24.8
pytesseract>=0.3.10
regex>=2024.5.15
rich>=13.7.1
sentencepiece>=0.2.0
PyYAML>=6.0.1
tokenizers>=0.19.1
torch>=2.3.0
tqdm>=4.66.4
transformers>=4.43.0
wandb>=0.17.0
unidecode>=1.3.8
